.archive/archive_pending/rakuten-trading-bot-v2/requirements.txt
fastapi
uvicorn
requests
feedparser
python-dotenv
google-cloud-firestore
google-cloud-storage
aiofiles
